Первая проверка ФЗ6
